1b/src/requirements.txt
# Core dependencies for sample scripts
requests>=2.31.0
google-generativeai>=0.3.0
google-cloud-texttospeech>=2.16.0
azure-cognitiveservices-speech>=1.34.0
openai>=1.3.0

# Optional local dependencies
# espeak (system package)
# festival (system package) 
# ffmpeg (system package)
